move DiffUtil callback for Movie into the adapter
